Move root logger to main.py and before aplication import to avoid circular dependency
