+++ .python-version
@@ -1,0 +1,4 @@
+3.6.12
+3.7.9
+3.8.6
+3.9.0rc1
